+++ typify/tests/schemas/simple-types.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct FloatsArentTerribleImTold {
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub flush_timeout: Option<f64>,
+    pub flush_timeout: Option<f32>,
 }
 impl From<&FloatsArentTerribleImTold> for FloatsArentTerribleImTold {
     fn from(value: &FloatsArentTerribleImTold) -> Self {
